pyproject.toml
[project]
name = "predictive-maintenance-system"
version = "1.0.0"
description = "Predictive maintenance for industrial machines using sensor data"
requires-python = ">=3.10"
dependencies = [
    "pandas==2.2.3",
    "numpy==1.26.4",
    "scikit-learn==1.5.2",
    "xgboost==2.1.3",
    "fastapi==0.115.6",
    "uvicorn==0.34.0",
    "pydantic==2.10.4",
    "pyyaml==6.0.2",
    "joblib==1.4.2",
    "httpx==0.28.1",
    "streamlit==1.41.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "ruff==0.8.6",
    "mypy==1.14.1",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
line-length = 99

[tool.mypy]
python_version = "3.10"
warn_return_any = true
